services/document-intelligence/requirements.txt
google-cloud-documentai==2.26.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.38.0
